tiffslide: limit zarr version in setup.cfg for python < 3.8
Close #25.

There's an incompatibility with zarr>=2.11.0 and tifffile<2022.3.25.
Since tifffile>=2022.2.2 dropped python 3.7 support in line with NEP29,
when installing manually or via pypi the default installation will
pull in an incompatible combination of zarr and tifffile on python=3.7

Note: this should only impact python 3.7 installs done via pypi
because tifffile>=2022.3.25 is available on conda-forge for 3.7.